+++ aidants_connect_web/tests/test_functional/test_homepage.py
@@ -1,0 +1,20 @@
+from django.contrib.staticfiles.testing import StaticLiveServerTestCase
+from selenium.webdriver.firefox.webdriver import WebDriver
+
+
+class HomePage(StaticLiveServerTestCase):
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        cls.selenium = WebDriver()
+        cls.selenium.implicitly_wait(10)
+        cls.selenium.get(f"{cls.live_server_url}/")
+
+    @classmethod
+    def tearDownClass(cls):
+        cls.selenium.quit()
+        super().tearDownClass()
+
+    def test_page_loads(self):
+        H1 = self.selenium.find_element_by_tag_name("h1")
+        self.assertEqual(H1.text, "Bienvenue sur Aidants Connect")
